fix admin + menu

+++ Sources/thembaihat.php
@@ -86,31 +86,47 @@ a:hover{
 	session_start();
 	include("./php/connect.php");
 	$sql=mysqli_query($con,"select * from baihat");
+
 	$row=mysqli_fetch_assoc($sql);
 		if(isset($_POST['up']))
 		{
+			
 			$tenbaihat=$_POST['tenbaihat'];
 			$casy=$_POST['casy'];
 			$theloai=$_POST['theloai'];
 			$file_name=$_FILES['upload']['name'];
+			$image=$_FILES['uploadimg']['name'];
 			$extent_file="mp3";
 			$pattern='#.+\.(mp3)$#i';
+			$pattern1='#\.(jpg|jpeg|gif|png)$#i';
 			
+			
+			if ($_FILES['upload']['size'] > 4000000)
+			{
 			if(preg_match($pattern,$file_name)==1)
 			{
-				if (isset($_POST['up']) && isset($_FILES['upload'])) {
-                    if ($_FILES['upload']['error'] > 0)
-                        echo "Upload lỗi rồi!";
-                    else {
-                        move_uploaded_file($_FILES['upload']['tmp_name'], 'data/' . $_FILES['upload']['name']);		
-                    }
-                }
-
-                $dest='data/'.$_FILES['upload']['name'];
-				if(file_exists($dest))
+				if(preg_match($pattern1,$image)==1)
 				{
+					
+					if (isset($_POST['up']) && isset($_FILES['upload']) && isset($_FILES['uploadimg'])) {
+						if ($_FILES['upload']['error'] > 0)
+							echo "Upload lỗi rồi!";
+						else {
+							move_uploaded_file($_FILES['upload']['tmp_name'], 'nhac/' . $_FILES['upload']['name']);	
+							move_uploaded_file($_FILES['uploadimg']['tmp_name'], 'image/' . $_FILES['uploadimg']['name']);
+							
+						}
+					}
+
+				$dest='nhac/'.$_FILES['upload']['name'];
+				$dest1='image/'.$_FILES['uploadimg']['name'];
+				if(file_exists($dest) && file_exists($dest1))
+				{	
+					$rtheloai=mysqli_query($con,"select id from theloai where tentheloai = '$theloai'");
+					$rowtl=mysqli_fetch_assoc($rtheloai);
+					$idtheloai = $rowtl['id'];
 					$flag=true;
-					$update=mysqli_query($con,"insert into baihat(tenbaihat,casy,theloai,duongdan) values('$tenbaihat','$casy','$theloai','$dest')");
+					$update=mysqli_query($con,"insert into baihat(tenbaihat,casy,theloai,duongdan,image,idtheloai) values('$tenbaihat','$casy','$theloai','$dest','$dest1','$idtheloai')");
 					if($update)
 					{
 						echo "Bài hát của bạn đã được đăng <a href='./admin/list_baihat.php'>Trở Lại</a><br />";
@@ -126,12 +142,26 @@ a:hover{
 					echo "Đăng nhạc thất bại!Trở về <a href='./admin/list_baihat.php'>Trở Lại</a><br />";
 				}
 
-			}
+				}
 			else
 			{
-				echo 'Sai định dạng file!';
+				
+				echo 'Sai định dạng file anh!';
 			}
+	
+		
+		
+		}else
+		{
+			echo 'Sai định dạng file nhac!';
 		}
+
+	}
+	else
+	{
+	echo "Dung luong anh qua lon";
+	}
+}
 			
 ?>
 <!DOCTYPE html>
@@ -161,12 +191,12 @@ a:hover{
 </head>
 <body>
     <div id="top">
-        <h3 style="color:#FFF";>Welcom Admin , <a href="../php/xulydangxuat.php" style="color:#FFF";>Logout</a></h3>
+		<h3 style="color:#FFF";>Welcom Admin , <a href="../php/xulydangxuat.php" style="color:#FFF";>Logout</a></h3>
     </div>
     <div id="menu">
         <ul>
             <li><a href="./admin/list_user.php">Quản Lý Thành Viên</a></li>
-            <li><a href="./adimn/list_chude.php">Quản Lý Chủ Đề</a></li>
+            <li><a href="./admin/list_chude.php">Quản Lý Chủ Đề</a></li>
             <li><a href="./admin/list_baihat.php">Quản Lý Bài Hát</a></li>
             <li><a href="./admin/list_comment.php">Quản Lý Bình Luận</a></li>
         </ul>
@@ -178,12 +208,13 @@ a:hover{
 	</div>
 	<div class="thongtin_dangky">
 		<div style="padding: 10px;">
-			<form name="form1" method="post" enctype="multipart/form-data" action="">
+			<form name="form1" method="post" enctype="multipart/form-data" action="thembaihat.php">
 				<table width="825" height="201" border="0">
 					<tr>
 						<td width="301" align="center" valign="top">
 						<p>
-							<input class="btup" type="file" name="upload" id="upload">
+							Nhạc: <input class="btup" type="file" name="upload" id="upload">
+							Ảnh :  <input class="btup" type="file" name="uploadimg" id="uploadimg" style="padding-right:0px;">
 						</p>
 						<table width="277" border="0">
 						  <tr>
@@ -198,14 +229,14 @@ a:hover{
 					      </tr>
 						  <tr>
 						    <td height="30" align="right"><b>Chủ Đề:</b></td>
-						    <td><label for="theloai"></label>
+						    <td><label for="chude"></label>
 						      <select name="theloai">
 										<?php
-											$chude=mysqli_query($con,"select * from chude");
-											while($row=mysqli_fetch_assoc($chude))
+											$theloai=mysqli_query($con,"select * from theloai");
+											while($row=mysqli_fetch_assoc($theloai))
 											{
 										?>
-											<option> <?php echo $row['chude']?></option>
+											<option> <?php echo $row['tentheloai']?></option>
 										<?php
 											}
 										?>
@@ -227,6 +258,8 @@ a:hover{
 					            <li> Không để tên bài hát có dấu.</li>
 					            <li> Không để tên ca sĩ trên bài hát.</li>
 					            <li> Chỉ up nhạc MP3.</li>
+								<li> Dung luong file anh khong vuot qua 4MB.</li>
+								<li> Chỉ up nhạc MP3.</li>
 			                </ul>
 							
 						</td>
